requirements.txt
Django==4.0
gunicorn==20.1.0
django-allauth==0.47.0
django-crispy-forms==1.12.0
django-recaptcha==3.0.0
dj-database-url==0.5.0
boto3==1.24.28
s3transfer==0.6.0
botocore==1.27.28
django-storages==1.11.1
psycopg2-binary==2.9.1
Pillow==9.2.0
pytz==2022.1
django-model-utils==4.2.0
django-address==0.2.8
celery==5.2.7
django-pwa==1.1.0
flower==1.2.0
redis==4.3.4